net: mctp: Allow limiting binds to a peer address
Prior to calling bind(), a program may call connect() on a socket to
restrict the bind to a specific peer address.

The smctp_type must match between connect() and bind() calls.

connect() is only used for bound sockets - send() is not available for
MCTP since a tag must be provided for each message.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>