net: mctp: Add support for gateway routing
This series adds a gateway route type for the MCTP core, allowing
non-local EIDs as the match for a route.

Example setup using the mctp tools:

    mctp route add 9 via mctpi2c0
    mctp neigh add 9 dev mctpi2c0 lladdr 0x1d
    mctp route add 10 gw 9

- will route packets to eid 10 through mctpi2c0, using a dest lladdr
of 0x1d (ie, that of the directly-attached eid 9).

The core change to support this is the introduction of a struct
mctp_dst, which represents the result of a route lookup. Since this
involves a bit of surgery through the routing code, we add a few tests
along the way.

We're introducing an ABI change in the new RTM_{NEW,GET,DEL}ROUTE
netlink formats, with the support for a RTA_GATEWAY attribute. Because
we need a network ID specified to fully-qualify a gateway EID, the
RTA_GATEWAY attribute carries the (net, eid) tuple in full:

    struct mctp_fq_addr {
        unsigned int net;
        mctp_eid_t eid;
    }

Of course, any questions, comments etc are most welcome.

To: Matt Johnston <matt@codeconstruct.com.au>
To: David S. Miller <davem@davemloft.net>
To: Eric Dumazet <edumazet@google.com>
To: Jakub Kicinski <kuba@kernel.org>
To: Paolo Abeni <pabeni@redhat.com>
To: Simon Horman <horms@kernel.org>
Cc: netdev@vger.kernel.org
Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>

---
Changes in v6:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v5: https://lore.kernel.org/r/20250702-dev-forwarding-v5-0-1468191da8a4@codeconstruct.com.au

Changes in v5:
- reinstate IFF_LOOPBACK check for new routes. Reported by Paolo Abeni
- mctp_sendmsg: return immediately on errors before we have allocated
  skb, preventing a mctp_dst_release() on uninit dst. Reported by Paolo
  Abeni.
- move mctp_dev_hold() to nlparse_populate, so we're consistent in the
  cleanup semantics. defer all stores to @rt to the success case.
- keep existing semantics of extaddr->halen checking
- Link to v4: https://lore.kernel.org/r/20250627-dev-forwarding-v4-0-72bb3cabc97c@codeconstruct.com.au

Changes in v4:
- avoid vla detection from const array-size initialiser, reported by ktr
- Link to v3: https://lore.kernel.org/r/20250625-dev-forwarding-v3-0-2061bd3013b3@codeconstruct.com.au

Changes in v3:
- 02/14 (new): Reduce frame size for route_input_cloned_frag kunit test,
  preventing -Wframe-size-larger-than warnings
- Link to v2: https://lore.kernel.org/r/20250619-dev-forwarding-v2-0-3f81801b06c2@codeconstruct.com.au

Changes in v2:
- 12/13: prevent uninitialsed gateway variable in nlparse_common
- 13/13: make test data static
- commit message spelling fixes
- Link to v1: https://lore.kernel.org/r/20250611-dev-forwarding-v1-0-6b69b1feb37f@codeconstruct.com.au

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 6,
    "change-id": "20250520-dev-forwarding-0711973470bf",
    "prefixes": [
      "net-next"
    ],
    "history": {
      "v1": [
        "20250611-dev-forwarding-v1-0-6b69b1feb37f@codeconstruct.com.au"
      ],
      "v2": [
        "20250619-dev-forwarding-v2-0-3f81801b06c2@codeconstruct.com.au"
      ],
      "v3": [
        "20250625-dev-forwarding-v3-0-2061bd3013b3@codeconstruct.com.au"
      ],
      "v4": [
        "20250627-dev-forwarding-v4-0-72bb3cabc97c@codeconstruct.com.au"
      ],
      "v5": [
        "20250702-dev-forwarding-v5-0-1468191da8a4@codeconstruct.com.au"
      ]
    }
  }
}