net: mctp: test: make cloned_frag buffers more appropriately-sized
In our input_cloned_frag test, we currently allocate our test buffers
arbitrarily-sized at 100 bytes.

We only expect to receive a max of 15 bytes from the socket, so reduce
to a more appropriate size. There are some upcoming changes to the
routing code which hit a frame-size limit on s390, so reduce the usage
before that lands.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>

---
v4:
- don't use const size_t as array size const, it isn't const enough, so
  triggers -Wvla-larger-than warnings. Reported by ktr.
v3:
- new commit, preventing -Wframe-larger-than issues for the upcoming
  change, reported by ktr.