net: mctp: usb: Add support for multi-packet transfers
The MCTP over USB spec allows us to pack multiple packets in one
transfer. Given the packet max length is 255, and the transfer max
length is 512, we can typically include two full-size packets per
urb submission.

To do this, we introduce a struct mctp_usb_tx, representing the ongoing
context for a tx. If sensible, a TX skb will be queued to the context
and the send deferred until the context is full, or the device queue
reports no more packets.

This typically requires a linear buffer for the 512-byte TX, which we
allocate along with the TX context. If the HCD supports scatter-gather
(without transfer boundary constraints), we can build a sg list instead,
making the context allocation much smaller.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>